Terminado a falta de cambiar la gestión de archivos y de Detalle_Pedido.cantidad_entregada que pasa a actualizarse con un trigger

+++ src/main/java/com/ajgestion/gestionpedidos/service/DetalleAlbaranService.java
@@ -1,10 +1,12 @@
 package com.ajgestion.gestionpedidos.service;
 
 import com.ajgestion.gestionpedidos.model.DetalleAlbaran;
+import com.ajgestion.gestionpedidos.model.DetallePedido;
 import com.ajgestion.gestionpedidos.repository.DetalleAlbaranRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -14,9 +16,11 @@ public class DetalleAlbaranService {
 
 
     private final DetalleAlbaranRepository detalleAlbaranRepository;
-
-    public DetalleAlbaranService(DetalleAlbaranRepository detalleAlbaranRepository){
+    private final DetallePedidoService detallePedidoService;
+    @Autowired
+    public DetalleAlbaranService(DetalleAlbaranRepository detalleAlbaranRepository, DetallePedidoService detallePedidoService){
         this.detalleAlbaranRepository = detalleAlbaranRepository;
+        this.detallePedidoService = detallePedidoService;
     }
 
     public List<DetalleAlbaran> obtenerTodos() {
@@ -31,7 +35,19 @@ public class DetalleAlbaranService {
         return detalleAlbaranRepository.save(detalleAlbaran);
     }
 
-    public void eliminar(Long id) throws EmptyResultDataAccessException {
-        detalleAlbaranRepository.deleteById(id);
+    @Transactional
+    public void eliminar(Long id) {
+        Optional<DetalleAlbaran> detalleAlbaran = detalleAlbaranRepository.findById(id);
+        if (detalleAlbaran.isPresent()) {
+            DetallePedido detallePedido = detalleAlbaran.get().getDetallePedido();
+            detallePedido.setCantidadEntregada(detallePedido.getCantidadEntregada() - detalleAlbaran.get().getCantidadEntregada());
+            detallePedidoService.guardar(detallePedido);
+            detalleAlbaranRepository.deleteById(id);
+        } else {
+            throw new EmptyResultDataAccessException(1);
+        }
+    }
+    public List<DetalleAlbaran> obtenerPorDetallePedido(Long id){
+        return detalleAlbaranRepository.findByDetallePedido_DetalleId(id);
     }
 }
